Add some extra comments inline

+++ router/admin.js
@@ -3,6 +3,7 @@ const router = express.Router();
 const { db, query } = require("../model/mysql");
 const { authenticateToken } = require('./auth');
 
+// Add new admin
 router.post("/", authenticateToken, (req, res) => {
     db.query(query.readAdmin(req.body.isbn), (err, results) => {
         if (err) throw err;
@@ -18,7 +19,11 @@ router.post("/", authenticateToken, (req, res) => {
     });
 });
 
+// Delete created admin
 router.delete("/", authenticateToken, (req, res) => {
+    if (req.body.email === 'admin@admin.com') {
+        return res.status(406).json({ Message: "You can't remove main admin" })
+    }
     db.query(query.deleteAdmin(req.body), (err) => {
         if (err) throw err;
         res.status(200).json({ Message: "Admin deleted successfully" })
